services/competitor-scan/main.py
from fastapi import FastAPI, HTTPException
from pymongo import MongoClient
from kafka import KafkaConsumer, KafkaProducer
from bs4 import BeautifulSoup
import requests, json, os, time, threading

app = FastAPI(title="Competitor Scan Service", version="1.0.0")

# ─── MONGODB ──────────────────────────────────────────
def get_db():
    client = MongoClient(os.getenv("MONGO_URI", "mongodb://mongo:27017/startup_validator"))
    return client.startup_validator

# ─── KAFKA PRODUCER ───────────────────────────────────
def get_producer():
    for i in range(10):
        try:
            return KafkaProducer(
                bootstrap_servers=os.getenv("KAFKA_BROKER", "kafka:9092"),
                value_serializer=lambda v: json.dumps(v).encode("utf-8")
            )
        except Exception as e:
            print(f"Producer attempt {i+1}/10: {e}")
            time.sleep(3)
    return None

# ─── COMPETITOR SEARCH ────────────────────────────────
def search_competitors(title: str, industry: str) -> dict:
    competitors = []
    saturation_score = 50

    try:
        # Search Product Hunt via web scrape
        query = f"{title} {industry} startup"
        headers = {"User-Agent": "Mozilla/5.0"}
        url = f"https://www.producthunt.com/search?q={query.replace(' ', '+')}"
        resp = requests.get(url, headers=headers, timeout=10)

        if resp.status_code == 200:
            soup = BeautifulSoup(resp.text, "html.parser")
            items = soup.find_all("h3", limit=5)
            for item in items:
                name = item.get_text(strip=True)
                if name and len(name) > 2:
                    competitors.append({
                        "name": name,
                        "source": "Product Hunt"
                    })

    except Exception as e:
        print(f"Scrape error: {e}")

    # Fallback: generate mock competitors if scraping fails
    if not competitors:
        mock_names = [
            f"{industry.title()}AI",
            f"Smart{title.split()[0].title()}",
            f"{industry.title()}Hub",
            f"Quick{title.split()[0].title()}",
            f"{industry.title()}Pro"
        ]
        competitors = [{"name": n, "source": "market research"} for n in mock_names[:3]]

    # Calculate saturation based on competitor count
    count = len(competitors)
    if count <= 2:
        saturation_score = 25
        saturation_level = "low"
    elif count <= 4:
        saturation_score = 55
        saturation_level = "medium"
    else:
        saturation_score = 80
        saturation_level = "high"

    return {
        "competitors": competitors,
        "competitor_count": count,
        "saturation_score": saturation_score,
        "saturation_level": saturation_level
    }

# ─── KAFKA CONSUMER THREAD ────────────────────────────
def consume_ideas():
    for i in range(15):
        try:
            consumer = KafkaConsumer(
                "idea-submitted",
                bootstrap_servers=os.getenv("KAFKA_BROKER", "kafka:9092"),
                value_deserializer=lambda m: json.loads(m.decode("utf-8")),
                group_id="competitor-scan-group",
                auto_offset_reset="earliest"
            )
            print("Competitor Scan: Kafka connected")
            break
        except Exception as e:
            print(f"Consumer attempt {i+1}/15: {e}")
            time.sleep(4)
            if i == 14:
                print("Kafka unavailable, consumer not started")
                return

    producer = get_producer()
    db = get_db()

    for message in consumer:
        idea     = message.value
        idea_id  = idea.get("idea_id")
        title    = idea.get("title", "startup")
        industry = idea.get("industry", "technology")

        print(f"Competitor Scan: processing {idea_id}")
        result_data = search_competitors(title, industry)

        result = {
            "idea_id":          idea_id,
            "competitors":      result_data["competitors"],
            "competitor_count": result_data["competitor_count"],
            "saturation_score": result_data["saturation_score"],
            "saturation_level": result_data["saturation_level"]
        }

        db.competitor_data.update_one(
            {"idea_id": idea_id},
            {"$set": result},
            upsert=True
        )

        if producer:
            producer.send("competitor-data-ready", result)
            producer.flush()

        print(f"Competitor Scan: done for {idea_id} → {result_data['saturation_level']} saturation")

@app.on_event("startup")
def startup_event():
    t = threading.Thread(target=consume_ideas, daemon=True)
    t.start()

# ─── ROUTES ───────────────────────────────────────────
@app.get("/health")
def health():
    return {"status": "ok", "service": "competitor-scan"}

@app.get("/competitors/{idea_id}")
def get_competitors(idea_id: str):
    db = get_db()
    data = db.competitor_data.find_one({"idea_id": idea_id})
    if not data:
        raise HTTPException(status_code=404, detail="Competitor data not found yet")
    data.pop("_id", None)
    return data

@app.post("/competitors/scan")
def scan_manual(payload: dict):
    title    = payload.get("title", "startup")
    industry = payload.get("industry", "technology")
    result   = search_competitors(title, industry)
    return {"title": title, "industry": industry, **result}